fix: divide GF into GFp and GFpn. #9
