+++ galois_field/__init__.py
@@ -1,3 +1,4 @@
-from galois_field.GFpn import GF
+from galois_field.GFpn import GFpn
+from galois_field.GFp import GFp
 from galois_field.ElementInGFp import ElementInGFp
 from galois_field.ElementInGFpn import ElementInGFpn
